Move converters to _core
