+++ parcels/_core/particleset.py
@@ -17,7 +17,7 @@ from parcels._core.kernel import Kernel
 from parcels.kernels import AdvectionRK4
 from parcels._core.particle import KernelParticle, Particle, create_particle_data
 from parcels._core.statuscodes import StatusCode
-from parcels.converters import _convert_to_flat_array
+from parcels._core.converters import _convert_to_flat_array
 
 __all__ = ["ParticleSet"]
